[OneExplorer] Move OneExplorer inside directory
This commit creates OneExplorer directory and moves the file into it
to expand the module.

ONE-vscode-DCO-1.0-Signed-off-by: Dayoung Lee <dayoung.lee@samsung.com>
